Exercises/Exercise-5/requirements.txt
pandas==2.2.3
psycopg2-binary==2.9.11
